+++ src/util/config-reader.ts
@@ -67,7 +67,6 @@ export class AppConfig {
                     const version = fs.readFileSync(versionFile, { encoding: 'utf-8', flag: 'r' });
                     if (version) {
                         AppConfig.reader.set('info.app.version', version.trim());
-                        log.info(`Application version ${version}`);
                     }
                 }
             } else {
@@ -123,6 +122,10 @@ export class AppConfig {
                     log.setLogFormat(LOG_FORMAT.COMPACT);
                 } 
             }
+            const version = AppConfig.reader.get('info.app.version');
+            if (version) {
+                log.info(`Application version ${version}`);
+            }      
             if (reloaded) {
                 log.info(`Configuration reloaded from ${reloadFile}`);
             } else if (errorInReload) {